tests/vboot: Move log screen tests to ui-broken-test.c
The developer, recovery and broken UIs all have the same log screens for
debug info and firmware log. Currently the main test cases for log
screens are put in the manual recovery test file (ui-manual-test.c) for
no particular reason. Those cases only appear in one UI test file
because we don't need to repeat the same tests for multiple UIs.

Considering that the log screens are more often used in the broken UI in
practice (because the debug info is usually the first thing to look at
when the device fails to boot up), move the test cases for log screens
from ui-manual-test.c to ui-broken-test.c. This will also simplify the
test cases by a little bit because in broken screens we don't have to
deal with the mocks for USB disks.

Still leave two basic tests for debug info and firmware log in
ui-manual-test.c as validity checks.

BUG=b:163301827
TEST=make unit-tests -j
BRANCH=none

Change-Id: I1eee9bd03489f78a83d009a1fa698babd9b2f161
Signed-off-by: Yu-Ping Wu <yupingso@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/5043271
Reviewed-by: Hsuan Ting Chen <roccochen@chromium.org>
Reviewed-by: Yidi Lin <yidilin@chromium.org>