tests/vboot/ui-developer-test: Remove power button shutdown test case
Similar to CL:5043271, all the firmware UIs share the same logic of
shutdown by power button press, so it is sufficient to have the shutdown
test case in one of the UI test files. Currently the same test case
appears in two files:

1. ui-developer-test.c: test_developer_ui_shutdown_timeout
2. ui-broken-test.c: test_broken_ui_power_button_shutdown

Since the former file is bigger and more complicated, keep the test case
only in ui-broken-test.c.

BUG=none
TEST=make unit-tests -j
BRANCH=none

Change-Id: I9a7809a65cdf8d8b68f1d1e71a29fb341424bed0
Signed-off-by: Yu-Ping Wu <yupingso@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/5090091
Reviewed-by: Yidi Lin <yidilin@chromium.org>