tests/drivers/input/mkbp/mkbp-test: Test holding a key on startup
Add a test case for the edge case of b/313812004, where the first event
AP sees on startup is an empty key_matrix. That happens for example when
powering off AP via selecting the "Power off" menu option by ENTER from
the firmware screen, and then powering on AP via EC console command
"powerbtn 200". In that case, when AP boots up, there will be 3 events
in EC's MKBP FIFO: "ENTER release", "POWER press" and "POWER release".

When the MKBP driver is registered by console_add_input_driver(), it
will flush the driver buffer before adding it to the console list. We
therefore expect EC's FIFO to be cleared with the code:

 while (havekey())
     getchar();

For the particular case of b/313812004 (fixed in CL:5171489), a single
call to getchar() should flush out all the 3 events in the FIFO, and
therefore if there are no more keys or buttons pressed, a subsequent
call to havekey() should return 0.

BUG=b:313812004
TEST=make tests/drivers/input/mkbp/mkbp-test
BRANCH=none

Change-Id: I8abdbb4cfc9fab11fbbf4d6eef7972da36d6ca0b
Signed-off-by: Yu-Ping Wu <yupingso@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/5208537
Reviewed-by: Hsuan Ting Chen <roccochen@chromium.org>
Reviewed-by: Yidi Lin <yidilin@chromium.org>