BRYA-ONLY: altfw: Fix segment type endianness
This patch fixes a bug that was introduced with CL:3379428 and later
fixed as part of CL:3461453. Brya branched between these patches. In
CL:3379428, we replaced the definitions from libpayload's <cbfs_core.h>
with the ones from <commonlib/bsd/cbfs_serialized.h>. Unfortunately the
payload segment type definitions (e.g. PAYLOAD_SEGMENT_CODE) have the
same names in both files, but were defined in different endianness, and
this discrepancy was missed.

CL:3461453 drive-by fixed this issue but is also a much larger patch
that switches to an entirely new CBFS API, something which we don't want
to cherry-pick back into a production branch. Therefore this patch only
extracts the relevant endianness fix.

This is probably not important enough to do a new firmware release for
by itself, but we should merge it onto the branch so it will piggyback
on the next release that's made for another reason.

BRANCH=brya
BUG=none
TEST=none

Change-Id: I335de057cebeb5d20d0de8204fed3511e0bdf39a
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/5337062
Reviewed-by: Subrata Banik <subratabanik@chromium.org>