vboot/Kconfig: Add SKIP_UI_DISPLAY for skipping firmware UI display
In early development, the display may not be ready. In that case, the
firmware UI is still functional (for example booting from USB in dev
mode), but the serial console will print out lots of error logs about
drawing failure, which are quite annoying.

Therefore, add a Kconfig option SKIP_UI_DISPLAY, which replaces the
usual ui_display() with a no-op function. This option should be enabled
only in development, and must not be left on for production.

BUG=b:336717577
TEST=emerge-geralt depthcharge  # with SKIP_UI_DISPLAY=y
BRANCH=none

Change-Id: I3598b61e390da2327ec866a728d93d31f86a28bd
Signed-off-by: Yu-Ping Wu <yupingso@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/5485922
Reviewed-by: Yidi Lin <yidilin@google.com>
Reviewed-by: Hsuan Ting Chen <roccochen@chromium.org>