BACKPORT: ec/cros: harmonize command timeouts
When commit 86398d3cb726b2608cf0ba1ceecdbdf58515f4e5 ("depthcharge:
increase EC erase timeout to 30 seconds") extended the erase command
timeout to 30 seconds, it didn't update the timeout in the underlying
send_packet function. Since send_packet's timeout comment says the
longer timeout it might use is for flash erase and 5 seconds was the old
maximum time to wait in send_command_proto3, move the constant
representing the current 30 second maximum and use it in both places so
they cannot diverge again.

It seems that without also changing the underlying function, an EC that
is running slowly can time out spuriously; this was observed on a system
in the automated test lab which is assumed to have an EC flash that's
been erased many times (so it might take longer to erase/program than
other systems).

BUG=b:333113770,b:222987250
TEST=EC software sync succeeds on the Yavilly with serial number
     5CD337NQXT, taking around 18 seconds to complete.
BRANCH=nissa

Change-Id: Ia409e2f4f498f867ce6ab90ed7ee04a6abaf4094
Signed-off-by: Peter Marheine <pmarheine@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/5500178
Reviewed-by: Ivan Chen <yulunchen@google.com>