rts5453: Correct return codes, adjust timeout
Return VB2_REQUEST_REBOOT_EC_TO_RO in error cases where we have trouble
contacting the PDC or resuming the EC PD stack. Rebooting the EC is the
best way to recover from these kinds of hangs.

Also adjust the polling loop at the end of the update that waits for
the PD stack to be re-initialized. Check less frequently (100ms->200ms)
to ease the strain on the EC PDC subsystem and expand the timeout from
5s to 7s since the PDC initialization sequence has gotten longer. There
is no impact on boot time because this polling loop only runs post-FWUP.

BUG=b:339511158
BRANCH=None
TEST=Flash to Brox

Change-Id: Iba269021b7a04ee580f023f8e757811434fb4468
Signed-off-by: Tristan Honscheid <honscheid@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/5525110
Reviewed-by: Keith Short <keithshort@chromium.org>
Reviewed-by: Bob Moragues <moragues@chromium.org>
Reviewed-by: Karthikeyan Ramasubramanian <kramasub@google.com>