dedede: Enable android GKI support on dedede
CONFIG_ANDROID_BOOT_DEVICES_DISK points to the internal MMC.
CONFIG_ANDROID_BOOT_DEVICES_REMOVABLE points to the USB controller.
Note that this was only tested on metaknight. Since both controllers are
embedded in the SoC the path should be the same on all jsl/dedede devices,
though this needs to be checked.

BUG=b:348346987
TEST=Boot android on metaknight

Change-Id: Ia0de6b73da2cdcaa859c57bcf4251d03b8e18953
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/5720977
Tested-by: Kornel Dulęba <korneld@google.com>
Reviewed-by: Jakub Czapiga <czapiga@google.com>
Commit-Queue: Kornel Dulęba <korneld@google.com>
Tested-by: Patryk Busse <pbusse@google.com>
Reviewed-by: Grzegorz Bernacki <bernacki@google.com>