nissa: Enable android GKI support on nissa
CONFIG_ANDROID_BOOT_DEVICES_DISK points to the internal MMC.
CONFIG_ANDROID_BOOT_DEVICES_REMOVABLE points to the USB controller.
Note that this was only tested on yavilly. Since both controllers are
embedded in the SoC the path should be the same on all ald-n/nissa devices,
though this needs to be checked.

BUG=b:352020523
TEST=Boot android on yavilly

Change-Id: Ife05e96568ad02fcab361fb52a3cfe4b92968d44
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/5806585
Tested-by: Łukasz Majczak <lmajczak@google.com>
Reviewed-by: Kornel Dulęba <korneld@google.com>
Commit-Queue: Łukasz Majczak <lmajczak@google.com>