rts5453: Enforce supported upgrade paths
Not every RTS5453 PDC FW version can be upgraded to every other. Add
a function that encodes the upgrade compatibility matrix and check it
before proceeding with the update. If an update is not supported, simply
skip it and print a warning. This prevents a boot loop due to wrong FW
continually being rejected by the PDC.

Add a unit test to verify the logic in test_rtk_upgrade_compatibility()

See linked bug for a spreadsheet with the compatibility matrix.

BUG=b:353510274
BRANCH=None
TEST=emerge-brox depthcharge
TEST=make tests/drivers/rts5453/rts5453-test

Change-Id: Ib6942e5bf2128940c37c298c06fb1db3b39cd867
Signed-off-by: Tristan Honscheid <honscheid@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/5714069
Reviewed-by: Karthikeyan Ramasubramanian <kramasub@google.com>
Reviewed-by: Keith Short <keithshort@chromium.org>