drivers/soc/cse: Remove unused function is_cse_fw_update_required
The function `is_cse_fw_update_required()` is not utilized in the code
path when late CSE sync is selected. This function was previously used
to determine if a CSE firmware update was necessary, but this logic is
no longer relevant in the current implementation.

BUG=b:305898363
TEST=Build successful for rex

Change-Id: Idc74c90ca6b73f0961c97a1aed013bb916454bed
Signed-off-by: Rishika Raj <rishikaraj@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/5759099
Reviewed-by: Subrata Banik <subratabanik@chromium.org>
Commit-Queue: Subrata Banik <subratabanik@chromium.org>
Reviewed-by: Dinesh Gehlot <digehlot@google.com>
Tested-by: Subrata Banik <subratabanik@chromium.org>