tests: Add stub functions for libpayload arch/cache.h
For unit tests, libpayload is built with CONFIG_LP_ARCH_MOCK=y, so cache
functions such as dcache_clean_all() defined in libpayload's
include/mock/arch/cache.h are used. For now, these functions are
implemented as empty macros, which are difficult to deal with when we
want to mock them in libpayload or depthcharge unit tests.

To allow easier mocking, we'd like to remove the default implementations
(empty macros) from libpayload's mock/arch/cache.h and replace them with
function declarations only, so that each test (either libpayload or
depthcharge) can control how to implement them.

To avoid co-dependent libpayload and depthcharge CLs, as a first step,
we add stub implementations in tests/stubs/libpayload/cache.c, and
include this file for all the tests with references to them. The
libpayload CL should be merged only after this CL is merged.

BUG=none
TEST=make unit-tests -j
BRANCH=none

Change-Id: I9e044052fe07c01490f5c567e110560dea5c6e45
Signed-off-by: Yu-Ping Wu <yupingso@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/5900702
Reviewed-by: Hsuan Ting Chen <roccochen@chromium.org>