fastboot: Add simple video output of fastboot
This is the most basic implementation of fastboot console. It is needed
just for early development and should be replaced with actual UI. It
displays only IP address. Default option for video as console backend is
not used, because scroll is too slow and it triggers timeouts on TCP
connections.

UART over servo is still functional.

BUG=b:370988331
TEST=FW_NAME=kano emerge-brya -j64 depthcharge
TEST=Run fastboot command and observe output on device's screen
BRANCH=main

Change-Id: I3cdd06514ae23a1d079e4e14f27b55b847ec1884
Signed-off-by: Tomasz Michalec <tmichalec@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/5906100
Reviewed-by: Jakub Czapiga <czapiga@google.com>