Makefile: Check depthcharge and libpayload architecture configs mismatch
This patch verifies that the same architecture configs are selected in
both depthcharge and libpayload. It throws a build error if there is
a mismatch.

For example, if CONFIG_ARCH_X86_64 is selected in depthcharge, it
verifies that the corresponding CONFIG_LP_ARCH_X86_64 is selected
in libpayload.

BUG=b:336088906
TEST=Able to build rex64 with correct configuration.
Verify build error by inducing config mismatch.
```
Makefile:111: *** CONFIG_ARCH_X86_64 is set but CONFIG_LP_ARCH_X86_64 is
not in /build/rex/tmp/portage/sys-boot/depthcharge-9999/work/build/rex64
/libpayload-rex64/libpayload.  Stop.
 * ERROR: sys-boot/depthcharge-9999::chromiumos failed (compile phase):
 *   emake failed

```

Change-Id: Ic1b50b182d5fe769f3b7dba0bdf1a35164687af0
Signed-off-by: Pranava Y N <pranavayn@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/5920299
Reviewed-by: Yu-Ping Wu <yupingso@chromium.org>
Reviewed-by: Subrata Banik <subratabanik@chromium.org>
Tested-by: Subrata Banik <subratabanik@chromium.org>