drivers/storage: Remove fill_write operation
Introduce blockdev_fill_write function which can be used in place of
fill_write callback. For most of the devices callback wasn't implemented
and it isn't possible to introduce significant optimizations.

Fastboot is only part of code that use this API.

Justification for 4MiB buffer size is carried over from mmc fill_write
implementation that was a base for the generic blockdev_fill_write
function.

BRANCH=main
BUG=b:378059710
TEST=Write super partition using fastboot

Change-Id: I45ccf99d921a3c749ca76d4bcd0fc7ba8539e773
Signed-off-by: Tomasz Michalec <tmichalec@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/6005079
Reviewed-by: Jan Dąbroś <dabros@google.com>