coreboot: Insert #address-cells and #size-cells for firmware node
Device nodes with children that contain address regions should have
the "#address-cells" and "#size-cells" properties.

The Linux Kernel recently enabled checking for these properties and
emitting big warnings [1] if they are missing. A patch was posted to
add the offending device into an exclusion list [2]. However the
firmware should be fixed to do the right thing.

[1] https://lore.kernel.org/all/Z0NUdoG17EwuCigT@sashalap/
[2] https://lore.kernel.org/all/20241125113151.107812-1-krzysztof.kozlowski@linaro.org/

BRANCH=none
BUG=b:380991384
TEST=Boot Linux and check /proc/device-tree/firmware has the two properties

Change-Id: I7ca66da3fa73c679ab66092ad8d612c2542e0f6b
Signed-off-by: Chen-Yu Tsai <wenst@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/6051580
Reviewed-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/6073650
Reviewed-by: Yidi Lin <yidilin@google.com>
Reviewed-by: Yu-Ping Wu <yupingso@chromium.org>