corsola: Enable pvmfw loading
This change enables pvmfw loading into system memory for corsola based
devices.

The change also includes an overwrite of the default pvmfw buffer address.
The default address points a place in memory suitable for x86_64
devices. We move the location of pvmfw next to depthcharge location in
order to avoid fragmentation of the memory. It is key, because the buffer
will occupy the memory until device reboot if pKVM has initialized
successfully.

BUG=b:354045389 b:359340876
TEST=Verify DTB node in sysfs
BRANCH=firmware-android-15949.B

Change-Id: If02fa4126d508fe1956ae0172c4e8203f521417a
Signed-off-by: Bartłomiej Grzesik <bgrzesik@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/6158202
Reviewed-by: Yu-Ping Wu <yupingso@chromium.org>