base/device_tree: Make END token part of struct_size
This is a copy of the same fix in CB:85462, for firmware branches that
are still using the depthcharge device tree implementation (i.e. before
CL:5655634).

According to the FDT specification the FDT_END token is supposed to be
the last token in the structure block, not a free-floating token
immediately outside of it. That means we're supposed to count it in
struct_size. It seems that the kernel never cared about this, but some
FDT parsing utilities like `fdtgrep` do.

BUG=b:381473816
TEST=cq
BRANCH=kukui, asurada, cherry, corsola, geralt

Change-Id: Ic282a6832c63761f375b9987f211e1ccd5472251
Signed-off-by: Yu-Ping Wu <yupingso@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/6073122
Commit-Queue: Yidi Lin <yidilin@google.com>
Reviewed-by: Yidi Lin <yidilin@google.com>
(cherry picked from commit 95b596adc06a2751e02d3a129cf77d403dab22c6)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/6072904