pvmfw: Add reserved memory entry with GSC boot params
This patch adds a new entry to the pvmfw configuration space with GSC
boot parameters. The entry holds the reserved memory blobs with
additional information. This information consists of UUID of the VM that
the reserved memory blob will copied into and DTB compatible string that
will be used to inform the VM about it's location in the guest memory.

BUG=b:354045389 b:359340876
TEST=pvmfw receives the correct data and is able to parse it
BRANCH=firmware-android-15949.B

Change-Id: I8fe41be985865473bbd134ab3a346c6648a2de16
Signed-off-by: Bartłomiej Grzesik <bgrzesik@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/6089574
Reviewed-by: Yu-Ping Wu <yupingso@chromium.org>