pvmfw: Add reserved memory entry with GSC boot params
This patch adds a new entry to the pvmfw configuration space with GSC
boot parameters. The entry holds the reserved memory blobs with
additional information. This information consists of name of the VM that
the reserved memory blob will copied into and DTB compatible string that
will be used to inform the VM about it's location in the guest memory.

BUG=b:410735713
TEST=pvmfw receives the correct data and is able to parse it
BRANCH=main

Change-Id: I0c1cb792d9c459a946cab7d661b19d0d7ee9e03e
Signed-off-by: Bartłomiej Grzesik <bgrzesik@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/6459286
Reviewed-by: Julius Werner <jwerner@chromium.org>
Reviewed-by: Yu-Ping Wu <yupingso@chromium.org>