Add support for COREBOOT_HEADERS variable
We are currently passing in the FW_CONFIG headers generated by coreboot
by copying the header into libpayload. This means we need to have a
unique libpayload for each different build.

This change adds a new variable we can set to pass in the location to
the coreboot headers so we can stop relying on the libpayload include
path.

BUG=b:392885541
BRANCH=none
TEST=emerge-brya depthcharge

Change-Id: I583f4d3f9cb9b01821e3a92f36db7077a05ea2b9
Signed-off-by: Raul E Rangel <rrangel@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/6219830
Reviewed-by: Jonathon Murphy <jpmurphy@google.com>
Commit-Queue: Jonathon Murphy <jpmurphy@google.com>