fastboot: Add support for "current-slot" variable
It returns the current a.k.a. active a/b slot the OS will boot from.
To avoid code duplication reuse "GptNextKernelEntry" from vboot to find
the active kernel partition. The same function is used by depthcharge
during boot. Note that this logic only works when android is installed
on internal storage, if the device is flashed with ChromeOS this command
will return an error.

based on:
commit c12e941f6ff5 ("fastboot: Add support for "current-slot" variable")

BRANCH=main
BUG=b:397322409
TEST=fastboot -s tcp:${dut_ip}:5554 getvar current-slot && \
     fastboot -s tcp:${dut_ip}:5554 set_active b && \
     fastboot -s tcp:${dut_ip}:5554 getvar current-slot && \
     fastboot -s tcp:${dut_ip}:5554 set_active a && \
     fastboot -s tcp:${dut_ip}:5554 getvar current-slot

Cq-Depend: chromium:6330154
Change-Id: Ie4969c6ca4fcc4548596c7e7782f7759b1a11088
Signed-off-by: Kornel Dulęba <korneld@google.com>
Signed-off-by: Tomasz Michalec <tmichalec@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/6574539
Reviewed-by: Julius Werner <jwerner@chromium.org>