fastboot: Decrease "max-download-size" to 16MB
Previously it was set to CONFIG_KERNEL_SIZE, which is equal to 512MB.
It was found that flashing big chunks of data to the internal storage
results in the OS image ending up corrupted. This is a temporary WA that
will be removed once b/397865136 is fixed. At that point we'll need to
figure out a proper default value, hence this commit shouldn't be simply
reverted.

BRANCH=none
BUG=b:397865136
TEST=Update super with "fastboot -s tcp:${dut_ip}:5554 flash super" and
     boot to OS.

Signed-off-by: Kornel Dulęba <korneld@google.com>
Change-Id: Iedc9d328e48e9f1a41be925e6a6a55d2f1963a16
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/6298609
Reviewed-by: Tomasz Michalec <tmichalec@google.com>