fastboot: Wrap fastboot run function
Fastboot is available only in dev build or production when
FASTBOOT_IN_PROD config is set. Because of that there are two different
entry functions for fastboot: dc_dev_fastboot() in dev build and
fastboot() in FASTBOOT_IN_PROD case.

Rename fastboot() to fastboot_run() and add fastboot() as a wrapper to
start fastboot in any build that has the fastboot available.

BRANCH=main
BUG=None
TEST=None

Change-Id: I969e53145f060c97e089363b54f8a43f18aa300c
Signed-off-by: Tomasz Michalec <tmichalec@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/6668479
Reviewed-by: Julius Werner <jwerner@chromium.org>