boot: Move devicetree update helpers separately
Move helper functions that update memory, chosen and commandline
devicetree nodes into a separate module. This module can be used either
when the devicetree is present along with kernel image in FIT format or
when the devicetree is present in separate disk partitions. As part of
this re-organization, following changes are made:
* Changed unsigned type to unsigned int type as suggested by repohook
  scripts
* Changed const char *[] to static const char *const [] as suggested by
  repohook scripts
* Kept the opening brace to EntryParams struct in the same line instead
  of newline
* Fixed a typo in one of the comments

BUG=b:394980221
BRANCH=None
TEST=USE=avb emerge-rauru coreboot libpayload vboot_reference
depthcharge chromeos-bootimage. Boot to ChromeOS.

Change-Id: I28fa538bd472e385899353e3c6c3ceb048f5b4ca
Signed-off-by: Karthikeyan Ramasubramanian <kramasub@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/6872730
Reviewed-by: Julius Werner <jwerner@chromium.org>
Reviewed-by: Yu-Ping Wu <yupingso@chromium.org>