board/bluey: Enable charging applet for low-battery charging
This commit introduces a new function, launch_charger_applet(), to
support off-mode and low-battery charging on the Bluey board.

The launch_charger_applet() function polls the EC for powerbtn event,
and if the AC adapter is present, it continues to charge
the battery. The charging loop is designed to:
 - Reboot: Exit the loop if the power button is pressed, allowing the
           system to boot normally during consecutive boot (tear-down
           ADSP-lite related instances while performing reboot).
 - Power-off:
     - Power off the system once the battery reaches full capacity.
     - Power off the system if the AC adapter is removed.

This feature is initiated from INIT_FUNC only when the system
boots in a low-battery state, as determined by a check against the EC.

TODO: Launch charger applet even for off-mode charging.

BUG=b:439819922
TEST=Able to put google/bluey into charging loop and charger till 100%.

Change-Id: Ic8ff046f954dab9c5d7ba3dde57970fa095d02e7
Signed-off-by: Subrata Banik <subratabanik@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/6880007
Reviewed-by: Kapil Porwal <kapilporwal@chromium.org>
Commit-Queue: Subrata Banik <subratabanik@chromium.org>
Auto-Submit: Subrata Banik <subratabanik@chromium.org>
Tested-by: Subrata Banik <subratabanik@chromium.org>