rauru: Reserve framebuffer memory in device tree
When a pre-allocated framebuffer is passed to the kernel for the
bootsplash, its memory region must be reserved in the device tree to
prevent the kernel from overwriting it.

This region must also be added to the reserved IOVA range to prevent
SMMU translation faults.

This patch adds a device tree fixup that reserves the framebuffer
memory and adds it to the IOVA reserved range when its physical
address is assigned.

BRANCH=none
BUG=b:435289727,b:319511268
TEST=Boot to ChromeOS desktop

Change-Id: I70a9a855ec18a5dd78371288052228d7751bcc23
Signed-off-by: Yidi Lin <yidilin@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/7040061
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Commit-Queue: Hung-Te Lin <hungte@chromium.org>