vboot: mocktpm: Mock pvmfw BootParam CBOR
This change adds the payload to the mocked secdata_get_pvmfw_params.

Prior to this change the secdata_get_pvmfw_params returned NULL and size
zero. This could lead to undesired behaviours and missing pvmfw for the
devices with mocktpm which is problematic.

With this change the secdata_get_pvmfw_params will returned a hard coded
mocked BootParam CBOR payload provided by @apronin that should be enough
to parse the CBOR and load pvmfw for the devices with mocktpm.

BUG=b:399081791
TEST=none
BRANCH=main

Change-Id: Id3709238f181179d428e5350366c2fde7a2aa3cd
Signed-off-by: Bartłomiej Grzesik <bgrzesik@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/6929663
Reviewed-by: Grzegorz Bernacki <bernacki@google.com>
Commit-Queue: Konrad Adamczyk <konrada@google.com>
Tested-by: Grzegorz Bernacki <bernacki@google.com>