board/bluey: Exit charging applet loop on power button press
The `launch_charger_applet` function now uses a `break` instead of
`cros_ec_ap_reset()` when the Power Button is pressed.

This samples the Power Button press and exits the infinite charging
loop, allowing the system to proceed with a normal boot to the OS.

The previous behavior of immediately triggering a reset is replaced
to facilitate a cleaner boot sequence when the user initiates a
power-on from the charging screen. A TODO is added to track the
requirement for an explicit device reset if needed for ADSP-lite
cleanup.

Additionally, drop the additional white-space after the
`DELAY_CHARGING_APPLET_MS` macro.

TEST=Able to bail-out charging loop upon pressing power-button.

Change-Id: Ieaab23fb96775a361232fc4be5b4f5d53a0b0854
Signed-off-by: Subrata Banik <subratabanik@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/7018611
Commit-Queue: Kapil Porwal <kapilporwal@chromium.org>
Tested-by: Subrata Banik <subratabanik@chromium.org>
Reviewed-by: Jayvik Desai <jayvik@google.com>
Reviewed-by: Kapil Porwal <kapilporwal@chromium.org>
Auto-Submit: Subrata Banik <subratabanik@chromium.org>