fastboot: add oem logs command
Add commands that dumps AP firmware logs to fastboot client.
Since sending the output through fastboot prints these messages
in the AP console, the output overwrites itself. To change that,
this commit adds parameter that decides if the message should also
be printed to console. All macros set this parameter as true by default
and the only use of false is for this oem command.
Output of the logs is sent line by line, to prevent splitting the
words or truncating the output.

BRANCH=main
BUG=b:449094080
TEST=fastboot -s tcp:X:Y oem logs
     multiple times and see all logs every time
TEST=make tests/fastboot/cmd-test

Change-Id: I7eda48e667874f3c3067e3846fdd71cf7c075022
Signed-off-by: Michał Barnaś <barnas@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/7030728
Reviewed-by: Tomasz Michalec <tmichalec@google.com>
Commit-Queue: Tomasz Michalec <tmichalec@google.com>