Introduce seed preparation for pKVM-DRNG
This change introduces seed preparation for pKVM-DRNG.

pKVM-DRNG is the pKVM vendor module for providing ARM TRNG
implementation for pVM on ARM devices that for some reason cannot
implement ARM TRNG in EL3/SEL1. In order for pKVM-DRNG to function
the initial seed has to be provided using DT.

The seed is generated using the previously introduced function
secdata_generate_randomness and placed in the remaining unused part
of the pvmfw buffer.

Then a DT node is added necessary for pkvm-drng to pickup the
generated seed. The new reserved-memory node is created to point to
the memory region where the seed had been placed, creating the
following structure:

/ {
    reserved-memory {
        pkvm_drng_seed: pkvm-drng-seed {
            compatible = "google,pkvm-drng";
            reg = <0x0 0xffb80000 0x0 0x80>;
            no-map;
        };
    };
};

The timestamps have been added to track the impact on boot time. On
rauru_navi 23 032 has been measured.

The entire implementation is gated by newly introduced
kconfig PKVM_DRNG_SEED. It is enabled by default on ARM devices with
pvmfw enabled.

BUG=b:449097147
BRANCH=main
TEST=The memory region is accessible to kernel and filled with random
    numbers

Change-Id: I951203c3267ddc1eab0ca89b158f7515b20f14b6
Signed-off-by: Bartłomiej Grzesik <bgrzesik@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/7003476
Reviewed-by: Julius Werner <jwerner@chromium.org>
Reviewed-by: Konrad Adamczyk <konrada@google.com>
Reviewed-by: Vadim Sukhomlinov <sukhomlinov@google.com>