boot: Add helper to get command line parameter values
Add `get_cmdline_uint32_value` to parse the kernel command line and
retrieve the value of a given parameter. This is useful for modules
that need to read specific parameters from the command line.

This function handles quoted values (e.g., `param="0x123"`). It
returns 0 on success and stores the value in the `out` parameter, or
-1 if the parameter is not found, the value is invalid, or the `out`
parameter is NULL.

Also, add unit tests for the new function.

BRANCH=main
BUG=b:438666196
TEST=COV=1 make tests/boot/commandline-test TEST_PRINT=1 -j

Change-Id: I5f38f7eddb178f742d0f9d169b8d894bfec10e15
Signed-off-by: Yidi Lin <yidilin@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/7104763
Reviewed-by: Yu-Ping Wu <yupingso@chromium.org>