drivers/storage/ufs: allow block access of data larger than 256MB
Previously the bounce buffer was created for the whole data buffer which
required allocation of more memory than we had. This commit changes
the behavior and allocates bounce buffer for every chunk with size of
max_transfer_blocks. This allows to provide buffer larger than 256MB
and will split the transfers to 16MB each, with proper handling of bounce
buffers if needed.

BRANCH=main
BUG=b:397865136
TEST=fastboot -s tcp:X:Y flash super super.img

Change-Id: Id298036b7f02d23bfce7a527af9e33917f62d63e
Signed-off-by: Michał Barnaś <barnas@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/7105542
Reviewed-by: Julius Werner <jwerner@chromium.org>