boot: Add support for minimum MTE control value
This change refactors the MTE command-line parsing to support a
minimum MTE control value, which can be used to enforce a baseline MTE
configuration.

The `apply_min_mte_ctrl` function is added to apply these minimum
settings, ensuring that the MTE configuration meets a certain
baseline.

This enhances the flexibility of MTE configuration by allowing the
enforcement of minimum MTE levels.

BRANCH=none
BUG=b:438666196
TEST=fastboot -s tcp:${DUT_IP}:5554 oem cmdline set bl.mtectrl=0x00053300
TEST=The FW log shows
     Newly appended kernel command line parameters:  kasan=on kasan.mode=sync kasan.fault=panic

Change-Id: Ie8ee8d10d50e0b66cdab9e74ee090cb8f0d40fde
Signed-off-by: Yidi Lin <yidilin@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/7123919
Reviewed-by: Yu-Ping Wu <yupingso@chromium.org>
Reviewed-by: Julius Werner <jwerner@chromium.org>