rauru: enable MTE for device with 16GiB DRAM
In current implementation, the MTE tag base address and size
are hardcoded in Kconfig, this is necessary because of a
limitation in coreboot where we are still figuring out a way
to get DRAM size at runtime, see b/461812409 for detail.

This also means: we need to disable MTE when the device's
DRAM size is not 16GiB until b/461812409 is fixed.

BRANCH=main
BUG=b:438666196
TEST=manual test
No-Presubmit: true

Change-Id: I4d2dcce8552a428f9297b7817db34730ea048808
Signed-off-by: Jeff Xu <jeffxu@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/7173749
Auto-Submit: Jeff Xu <jeffxu@google.com>
Tested-by: Jeff Xu <jeffxu@google.com>
Commit-Queue: Jeff Xu <jeffxu@google.com>
Reviewed-by: Yu-Ping Wu <yupingso@chromium.org>