sound/hda_codec: Add amp mute option for hda codec driver
On some devices, a "pop" noise occurs during reboot or shutdown if
audio was playing just before the restart. This is caused by the
speaker amplifier being active during the power transition. A workaround
is to mute the speaker during boot.

This change introduces a new `mute_amp_nid` option to the HDA codec
driver. When a node ID is provided via this option, the driver will
temporarily unmute the amplifier when the `devbeep` starts and mute it
again when it stops.

This ensures the `devbeep` remains audible while still preventing the
unwanted pop noise during system restarts. The option is enabled for
the ALC256 codec with NID 0x14.

BUG=b:439638686
BRANCH=main
TEST=play video and reboot on caboc

Change-Id: Ifa8a728205665774bf7a07ac89551342e4195ff7
Signed-off-by: Terry Cheong <htcheong@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/7197740
Reviewed-by: Yidi Lin <yidilin@google.com>