board/bluey: Enable heartbeat when not charging in off-mode
In off-mode charging, when the battery is not drawing current, it may
be fully charged or battery is disconnected. In this scenario, the EC
is instructed to perform a heartbeat discharge wake if the external
power is connected.

This is achieved by calling the recently added
`cros_ec_enable_offmode_heartbeat` function which leverages the
`EC_CMD_ENABLE_OFFMODE_HEARTBEAT` host command. This ensures the AP
will be woken up after the battery discharges by 1% if the external
power is still  connected.

BRANCH=none
BUG=b:439819922
TEST=Verify off-mode charging behavior on Google/Quenbi,
specifically when the battery is fully charged.

Cq-Depend: chromium:7201384, chromium:7180720
Change-Id: If549ca150a295f52372c9be33a48f4e76069569b
Signed-off-by: Jayvik Desai <jayvik@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/7207894
Reviewed-by: Kapil Porwal <kapilporwal@chromium.org>
Reviewed-by: Subrata Banik <subratabanik@chromium.org>
Commit-Queue: ChromeOS Auto Retry <chromeos-auto-retry@chromeos-bot.iam.gserviceaccount.com>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>