base/gpt: Erase first 256MiB of partition
When erase method is not implemented for the block storage, fill write
only the first 256MiB of the partition. This workaround is required,
because some fastboot flows try to erase large partitions and doing fill
write over +100GB is time consuming.

This workaround should be evaluated after implementing poper erase
method for the drivers that are in use.

BRANCH=main
BUG=b:396352272
TEST=fastboot -s tcp:${dut_ip}:5554 -w

Change-Id: Ic51e127dac72c1852c562f138f1059eef8448d83
Signed-off-by: Tomasz Michalec <tmichalec@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/7221861
Reviewed-by: Konrad Adamczyk <konrada@google.com>