board/bluey: Use standard reboot to exit charging applet
Now that warm-reset functionality is stable on the platform, switch
from the EC-based reboot workaround to the standard reboot() call
when exiting the charging applet.

Previously, a workaround (cros_ec_reboot) was implemented because the
system-level warm reset was not successfully transitioning to the OS
boot flow. Using the standard reboot() ensures a cleaner handover
as per the platform's power-on logic.

Resolves the workaround tracked in b/429523381.

BUG=b:429523381
BRANCH=None
TEST=Verified on bluey that pressing the power button during the
charging applet triggers a standard reboot and successfully boots
into the OS.

Change-Id: Ib87e2187baf5a15d4ec3618c82544b58ddd3b9f5
Signed-off-by: Subrata Banik <subratabanik@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/7267790
Tested-by: Kapil Porwal <kapilporwal@chromium.org>
Reviewed-by: Kapil Porwal <kapilporwal@chromium.org>
Commit-Queue: Kapil Porwal <kapilporwal@chromium.org>