boot: Disable MTE if tag storage is not reserved
When MTE is enabled, a dedicated memory region (tag storage) must be
reserved for memory tags.

This commit introduces a check in android_mte_setup to verify that the
tag storage exists before enabling MTE in the kernel command line. If
the storage is not found, MTE is disabled to prevent potential issues.

Additionally, if MTE is disabled, this change ensures that any reserved
tag storage is freed, making the memory available for other uses.

BRANCH=none
BUG=b:438666196
TEST=The memory region for tag storage is reserved when MTE is enabled.
TEST=The tag storage is released when MTE is disabled.

Change-Id: Id4751332016b360f1a38e4adcd8bdc9c4afbaf9b
Signed-off-by: Yidi Lin <yidilin@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/7366319
Reviewed-by: Yu-Ping Wu <yupingso@chromium.org>
Reviewed-by: Julius Werner <jwerner@chromium.org>